MP2-575 handle lead aprtner replacement + fix max 1 lead partner

+++ ems-api/src/main/kotlin/io/cloudflight/ems/api/project/dto/OutputProjectPartner.kt
@@ -4,5 +4,5 @@ data class OutputProjectPartner (
     val id: Long?,
     val name: String,
     val role: ProjectPartnerRole,
-    val sortNumber: Long? = null
+    val sortNumber: Int? = null
 )
