+++ ems-server/src/main/kotlin/io/cloudflight/ems/project/entity/ProjectPartner.kt
@@ -30,7 +30,7 @@ data class ProjectPartner(
     val role: ProjectPartnerRole,
 
     @Column
-    val sortNumber: Long? = null
+    val sortNumber: Int? = null
 
 ) {
     override fun toString(): String {
